feat: Improved installation via metadata
closes #1

+++ taku/__init__.py
@@ -1,5 +1,6 @@
 import argparse
 import os
+import platform
 import shutil
 import stat
 import subprocess
@@ -23,9 +24,12 @@ from .exceptions import TemplateNotFoundError
 from .run import run_script, _resolve_script, default_scripts_dir
 
 try:
-    from rich_argparse import RichHelpFormatter
+    from rich_argparse import ArgumentDefaultsRichHelpFormatter
 
-    formatter_class = RichHelpFormatter
+    # from rich import print as rprint
+
+    formatter_class = ArgumentDefaultsRichHelpFormatter
+    # print = lambda *args, **kwargs: rprint(*args, **kwargs)  # type: ignore
 except ImportError:
     formatter_class = argparse.ArgumentDefaultsHelpFormatter
 
@@ -64,7 +68,7 @@ def _list_scripts(scripts: Path) -> list[str]:
 cmd("run")(run_script)
 
 
-@cmd("list", aliases=["ls"])
+@cmd("list", aliases=["ls"], formatter_class=formatter_class)
 def list_scripts(
     scripts: Annotated[Path, ArgSpec(ignore=True)],
     templates: Annotated[
@@ -86,7 +90,7 @@ def list_scripts(
     print("\n".join(scripts_list))
 
 
-@cmd("new")
+@cmd("new", formatter_class=formatter_class)
 def new_script(
     scripts: Annotated[Path, ArgSpec(ignore=True)],
     name: Annotated[str, ArgSpec(help="Name of the new script")],
@@ -134,7 +138,7 @@ def new_script(
     print(f"script {name} created")
 
 
-@cmd("get")
+@cmd("get", formatter_class=formatter_class)
 def get_script(
     scripts: Annotated[Path, ArgSpec(ignore=True)],
     name: Annotated[str, ArgSpec(help="Name of the script")],
@@ -161,7 +165,7 @@ def get_script(
         print(key, ":", value)
 
 
-@cmd("rm")
+@cmd("rm", formatter_class=formatter_class)
 def rm_script(
     scripts: Annotated[Path, ArgSpec(ignore=True)],
     name: Annotated[str, ArgSpec(help="Name of the script")],
@@ -175,7 +179,7 @@ def rm_script(
     push_scripts(scripts)
 
 
-@cmd("edit")
+@cmd("edit", formatter_class=formatter_class)
 def edit_script(
     scripts: Annotated[Path, ArgSpec(ignore=True)],
     name: Annotated[str, ArgSpec(help="Name of the script to edit")],
@@ -191,7 +195,7 @@ def edit_script(
     push_scripts(scripts)
 
 
-@cmd("install")
+@cmd("install", formatter_class=formatter_class)
 def install_scripts(
     scripts: Annotated[Path, ArgSpec(ignore=True)],
     name: Annotated[
@@ -205,9 +209,14 @@ def install_scripts(
             help="Optional name when installed in PATH, only used when installing a single script"
         ),
     ] = None,
+    target_dir: Annotated[
+        Path,
+        "--target-dir",
+        "-t",
+        ArgSpec(help="Target directory, should be in the PATH"),
+    ] = Path.home() / ".local/bin",
 ):
-    """Install a script to ~/.local/bin"""
-    target_dir = Path.home() / ".local/bin"
+    """Install a script to the specified target directory"""
     target_dir.mkdir(parents=True, exist_ok=True)
 
     install_name = install_name or name
@@ -221,7 +230,7 @@ def install_scripts(
         else {s: s for s in _list_scripts(scripts)}
     )
     exec_path = Path(sys.executable).parent / "tax"
-
+    host = platform.node()
     for script_name, script_install_name in to_install.items():
         target_file = target_dir / script_install_name
         metadata_file = scripts / script_name / "meta.toml"
@@ -245,35 +254,47 @@ exec {exec_path} "{script_name}" "$@"
         metadata = (
             tomllib.loads(metadata_file.read_text()) if metadata_file.exists() else {}
         )
-        metadata["install_name"] = script_install_name
+        metadata[host] = {
+            "install_name": script_install_name,
+            "target_dir": str(target_dir),
+        }
         metadata_file.write_text(tomli_w.dumps(metadata))
         print(f"Installed {script_name} to {target_file}")
 
 
-@cmd("uninstall")
+@cmd("uninstall", formatter_class=formatter_class)
 def uninstall_scripts(
     scripts: Annotated[Path, ArgSpec(ignore=True)],
     name: Annotated[
         str, ArgSpec(help="Name of the script to uninstall, use 'all' for all scripts")
     ],
 ):
-    """Uninstall a script from ~/.local/bin"""
-    target_dir = Path.home() / ".local/bin"
+    """Uninstall a script using its metadata"""
 
     if name != "all":
         _resolve_script(scripts, name)
 
     to_uninstall = [name] if name != "all" else _list_scripts(scripts)
-
+    host = platform.node()
     for script_name in to_uninstall:
         metadata_file = scripts / script_name / "meta.toml"
-        metadata = (
-            tomllib.loads(metadata_file.read_text()) if metadata_file.exists() else {}
-        )
-        target_file = target_dir / metadata.get("install_name", script_name)
+        if not metadata_file.exists():
+            print(f"Skipping {script_name}, no metadata file found")
+            continue
+
+        metadata = tomllib.loads(metadata_file.read_text())
+
+        # Check if host has metadata
+        if host not in metadata:
+            print(f"No installation metadata found for {script_name} on host {host}")
+            continue
+
+        host_metadata = metadata[host]
+        target_dir = Path(host_metadata["target_dir"])
+        target_file = target_dir / host_metadata["install_name"]
 
         if target_file.exists():
-            metadata.pop("install_name", None)
+            metadata.pop(host, None)
             metadata_file.write_text(tomli_w.dumps(metadata))
             target_file.unlink()
             print(f"Uninstalled {script_name} from {target_file}")
